+++ src/libsac2c/codegen/icm_comment.c
@@ -42,7 +42,7 @@
  */
 #define ICM_VARANY(dim, name)                                                            \
     {                                                                                    \
-        size_t i;                                                                           \
+        int i;                                                                           \
         for (i = 0; i < dim; i++) {                                                      \
             ICM_ANY (name[i])                                                            \
         }                                                                                \
@@ -50,7 +50,7 @@
 
 #define ICM_VARNT(dim, name)                                                             \
     {                                                                                    \
-        size_t i;                                                                           \
+        int i;                                                                           \
         for (i = 0; i < dim; i++) {                                                      \
             ICM_NT (name[i])                                                             \
         }                                                                                \
@@ -58,7 +58,7 @@
 
 #define ICM_VARID(dim, name)                                                             \
     {                                                                                    \
-        size_t i;                                                                           \
+        int i;                                                                           \
         for (i = 0; i < dim; i++) {                                                      \
             ICM_ID (name[i])                                                             \
         }                                                                                \
@@ -66,7 +66,7 @@
 
 #define ICM_VARINT(dim, name)                                                            \
     {                                                                                    \
-        size_t i;                                                                           \
+        int i;                                                                           \
         for (i = 0; i < dim; i++) {                                                      \
             ICM_INT (name[i])                                                            \
         }                                                                                \
